frontend/requirements.txt
flask
python-dotenv
flask-restful
requests
flask-login
flask_wtf
email_validator
PyJWT
